File renaming and settings component
