+++ frontend/src/pages/Profiles_Page/ProfileInfo.tsx
@@ -3,7 +3,8 @@ import IonIcon from '@reacticons/ionicons';
 import { ReactElement } from 'react';
 import { Profile, UserStatus, Roles } from './Interface';
 import Sidebar from '../Sidebar_Components/Sidebar'
-import NotifBar from '../components/notification_bar.js';
+import NotifBar from './components/Notification_Bar';
+import SettingsBar from './components/Settings';
 import React, { useState } from 'react';
 
 interface ProfileInformation {
@@ -40,20 +41,7 @@ const ProfileInfo = (props: ProfileInformation): ReactElement => {
                 {/*Settings Div 
                     - output will depend on main role
                 */}
-                <div className='flex flex-col w-3/4 md:w-3/5 mt-[5%] text-4xl text-[#262B6C]'>
-                    <button className='border-solid border-t border-[#262B6C] text-left flex-row inline-flex px-8'>
-                        <p className='py-16 w-3/4 h-full'>Personal Stats</p>
-                        <div className='py-16 w-1/4 h-full text-right'><IonIcon name="chevron-forward-outline" /></div>
-                    </button>
-                    <button className='border-solid border-t border-[#262B6C] text-left flex-row inline-flex px-8'>
-                        <p className='py-16 w-3/4 h-full'>Meeting Stats</p>
-                        <div className='py-16 w-1/4 h-full text-right'><IonIcon name="chevron-forward-outline" /></div>
-                    </button>
-                    <button className='border-solid border-t border-[#262B6C] text-left flex-row inline-flex px-8'>
-                        <p className='py-16 w-3/4 h-full'>Roles</p>
-                        <div className='py-16 w-1/4 h-full text-right'><IonIcon name="chevron-forward-outline" /></div>
-                    </button>
-                </div>
+                <SettingsBar />
             </div>
             <NotifBar />
         </div>
